feat(controller): add `RedirectView` and flash attributes to the methods of FileController

+++ src/main/java/com/file/storage/controller/FileController.java
@@ -4,13 +4,22 @@ import com.file.storage.dto.file.FileDeleteRequest;
 import com.file.storage.dto.file.FileDownloadRequest;
 import com.file.storage.dto.file.FileRenameRequest;
 import com.file.storage.dto.file.FileUploadRequest;
+import com.file.storage.exception.file.InvalidFileDeleteRequestException;
+import com.file.storage.exception.file.InvalidFileDownloadRequestException;
+import com.file.storage.exception.file.InvalidFileRenameRequestException;
+import com.file.storage.exception.file.InvalidFileUploadRequestException;
 import com.file.storage.service.FileService;
+import com.file.storage.util.ValidationUtils;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.view.RedirectView;
 
 @Controller
 @RequestMapping("/files")
@@ -21,7 +30,13 @@ public class FileController {
 
     @GetMapping(produces = MediaType.APPLICATION_OCTET_STREAM_VALUE)
     @ResponseBody
-    public ResponseEntity<ByteArrayResource> downloadFile(@ModelAttribute("fileDownloadRequest") FileDownloadRequest fileDownloadRequest) {
+    public ResponseEntity<ByteArrayResource> downloadFile(@Valid @ModelAttribute("fileDownloadRequest") FileDownloadRequest fileDownloadRequest,
+                                                          BindingResult bindingResult) {
+
+        if (bindingResult.hasErrors()) {
+            throw new InvalidFileDownloadRequestException(ValidationUtils.getErrorMessage(bindingResult));
+        }
+
         ByteArrayResource file = fileService.downloadFile(fileDownloadRequest);
         return ResponseEntity.ok()
                 .header("Content-Disposition", "attachment; filename=" + fileDownloadRequest.getName())
@@ -29,20 +44,48 @@ public class FileController {
     }
 
     @PostMapping
-    public String uploadFile(@ModelAttribute("fileUploadRequest") FileUploadRequest fileUploadRequest) {
+    public RedirectView uploadFile(@Valid @ModelAttribute("fileUploadRequest") FileUploadRequest fileUploadRequest,
+                                   BindingResult bindingResult,
+                                   RedirectAttributes redirectAttributes) {
+
+        if (bindingResult.hasErrors()) {
+            throw new InvalidFileUploadRequestException(ValidationUtils.getErrorMessage(bindingResult));
+        }
+
         fileService.uploadFile(fileUploadRequest);
-        return "redirect:/";
+
+
+        redirectAttributes.addFlashAttribute("success", "File uploaded successfully");
+        return new RedirectView("/");
     }
 
     @PutMapping
-    public String renameFile(@ModelAttribute("fileRenameRequest") FileRenameRequest fileRenameRequest) {
+    public RedirectView renameFile(@Valid @ModelAttribute("fileRenameRequest") FileRenameRequest fileRenameRequest,
+                                   BindingResult bindingResult,
+                                   RedirectAttributes redirectAttributes) {
+
+        if (bindingResult.hasErrors()) {
+            throw new InvalidFileRenameRequestException(ValidationUtils.getErrorMessage(bindingResult));
+        }
+
         fileService.renameFile(fileRenameRequest);
-        return "redirect:/";
+
+        redirectAttributes.addFlashAttribute("success", "File renamed successfully");
+        return new RedirectView("/");
     }
 
     @DeleteMapping
-    public String deleteFile(@ModelAttribute("fileDeleteRequest") FileDeleteRequest fileDeleteRequest) {
+    public RedirectView deleteFile(@Valid @ModelAttribute("fileDeleteRequest") FileDeleteRequest fileDeleteRequest,
+                                   BindingResult bindingResult,
+                                   RedirectAttributes redirectAttributes) {
+
+        if (bindingResult.hasErrors()) {
+            throw new InvalidFileDeleteRequestException(ValidationUtils.getErrorMessage(bindingResult));
+        }
+
+        redirectAttributes.addFlashAttribute("success", "File deleted successfully");
         fileService.deleteFile(fileDeleteRequest);
-        return "redirect:/";
+
+        return new RedirectView("/");
     }
 }
